pyproject.toml
[tool.poetry]
name = "django-weather"
version = "0.1.0"
description = ""
authors = ["Yuriy Peskov <yuriy.peskov@gmail.com>"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.12"
Django = "^5.0.2"
psycopg = "^3.1.18"
djangorestframework = "^3.14.0"
icecream = "^2.1.3"
djangorestframework-simplejwt = "^5.3.1"
pytest-django = "^4.8.0"
pytest-cov = "^4.1.0"
requests = "^2.31.0"
celery = "^5.3.6"
redis = "^5.0.2"
django-celery-beat = "^2.6.0"
django-celery-results = "^2.5.1"
watchdog = "^4.0.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

